Issue #354: Perform optional cleanup after image upload
Having both master and dev build jobs on the same machine
takes disk space faster and make it less clear
which images are safe to delete.

This inptroduces cleanup flag which causes the upload script
to delete all images with the selected tag,
that catches also images not being uploaded, such as dev-contiv-plugins.