Add unit tests for handling of transient Etcd errors (#572)
This pull request add unit tests for testing the handling of transient Etcd errors. The unit tests have been added to the policy_reflector_test, which contains data store resync tests for the whole KSR.

New tests:

testResyncPolicyTransientDsError: is called from the main test function that initializes the Policy reflector. The test first initializes test data in the data store, and get a snapshot of Policy Reflector's stats. Then, Etcd is checked a couple of times (checkEtcdTransientError) to properly initialize etcdMonitor. Finally, a transient data store error is emulated by clearing the data store and etcd is check again (checkEtcdTransientError), which should cause a data store resync. The test then checks that the data store resync has been properly executed.
Modified tests:

testResyncPolicyDataStoreDownThenAdd: uses the etcd monitor instead of KSR event functions to generate data store up/down events. This way, this test also covers a couple of paths in etcdMonitor.processEtcdMonitorEvent
Minor cleanup:

in all tests, ***Writers have been renamed to ***Brokers
minor naming cleanups
typo fixes
The mockKeyProtoValBroker's GetValue function has been modified to more closely mimic the real Etcd Broker's GetValue function because the GetValue function is used in the transient error detection mechanism. This change means that if a value for a given key is not found, the function now returns found=false and error=nil (not finding a value for a key is not an error condition). All tests using mockKeyProtoValBroker's had to be modified to work with the new logic.