Add support for configurable default route (default GW)  (#575)
This PR removes the hardcoded default route from VPP to the host stack. Instead,
it allows specifying a custom default gateway as a node-config option Gateway
in the contiv-vpp.yaml deployment file.

In order to remove the default route to host on VPP, specific routes need to be added:
VPP needs to be able to reach host's local IP addresses (addresses on interfaces not bound to VPP),
since one of them may be the main IP address of the node, as k8s sees it. Since this IP is used
for NAT translations of services with host networking, PODs behind VPP need to be able to reach it.

Note: "github.com/vishvananda/netlink" should be added as an explicit dependency once #574 is merged. Not required for successful build, since it already exists in the vendor folder due to transient dependencies.