SNAT traffic leaving cluster network.
This commit implements source NATing for all traffic leaving the cluster
network, which in effect opens up the Internet access for all pods.
The SNAT was included into the Service plugin in order to keep the NAT-related
configuration all in one place.

The solution is to add the IP address of the default GW interface into
the pool of VPP/NAT44 addresses and to enable postrouting on that
interface.

The traffic going between cluster nodes should not be NATed otherwise
the ACLs of the destination node would no longer match against
pod IPs, but rather against node IPs, which breaks the semantic.
It is possible to separate external traffic from the internal one
only with the assistance of VXLANs, therefore the SNAT is not supported
and gets disabled in the L2-only mode.