Renderer cache API - cache unified for both ACL and VPPTCP renderer
RendererCache combines capabilites of the VPPTCP and ACL caches
under a unified interface.

The rules are grouped into tables (ContivRuleTable type) and the
configuration is represented as a list of local tables, applied
on the ingress or the egress side of pods, and a single global table,
applied on the interfaces connecting the node with the rest
of the cluster.

The list of local tables is minimalistic in the sense that pods with
the same set of rules will share the same local table. Whether shared
tables are installed in one instance or as separate copies for each
associated pod is up to the renderer (usually determined by
the capabilities of the destination network stack).

All tables match only one side of the traffic - either ingress or
egress, depending on the cache orientation as selected in the Init method.
The cache combines the received ingress and egress Contiv rules
into the single chosen direction in a way that maintains the original
semantic (the global table is introduced to accomplish the task).

The rules are ordered in tables such that if rule *r1* matches subset
of the traffic matched by *r2*, then r1 precedes r2 in the list.
It is the order at which the rules should by applied by the rule
matching algorithm in the destination network stack (otherwise the
more specific rules could be overshadowed and never matched).

There are two types of tables distinguished:
  1. Local table: should be applied to match against traffic leaving
                  (IngressOrientation) or entering (EgressOrientation)
                  a selected subset of pods.
                  Every pod has at most one local table installed at
                  any given time. For a given local table, the set
                  of rules is immutable. Different content is treated
                  as a new local table (and the original table may
                  get unassigned from some or all originally
                  associated pods).
                  Local table has always at least one rule, otherwise
                  it is simply not tracked and returned by the cache.
  2. Global table: should be applied to match against traffic entering
                   (IngressOrientation) or leaving (EgressOrientation)
                   the node. There is always exactly one global table
                   installed (per node).
                   The global table may contain an empty set of rules
                   (meaning ALLOW-ALL).