Secure VPP debug access (#597)
The current method of listening on a TCP socket allows non-root
users on the local host, and remote users to alter the network
configuration.  By switching this to a UNIX socket that only root
can access on the local host, we remove this vulnerability.

Note:  The documented netcat client does not support tab completion.
In the future, we should be able to make a simple CLI program that
will work with tab completion.

Drive by:  Cleaned up an inline bash script so it now reports error
status correctly, and made it more readable.