[WIP, READY FOR REVIEW] Integration of policies with services and the Internet access (#609)
* Keep kube-system pods accessible regardless of the policies installed.

Pods should be able to access kubernetes services (e.g. DNS)
even if they are isolated from the kube-system namespace by the
installed K8s network policies.
However, this is not the case in the opposite direction.
Policy may disallow kube-system pod to conntact pod from another
namespace.

* Update NAT binary APIs.

* Expose default GW IP via Contiv plugin API.

* SNAT traffic leaving cluster network.

This commit implements source NATing for all traffic leaving the cluster
network, which in effect opens up the Internet access for all pods.
The SNAT was included into the Service plugin in order to keep the NAT-related
configuration all in one place.

The solution is to add the IP address of the default GW interface into
the pool of VPP/NAT44 addresses and to enable postrouting on that
interface.

The traffic going between cluster nodes should not be NATed otherwise
the ACLs of the destination node would no longer match against
pod IPs, but rather against node IPs, which breaks the semantic.
It is possible to separate external traffic from the internal one
only with the assistance of VXLANs, therefore the SNAT is not supported
and gets disabled in the L2-only mode.

* Add API to the Contiv plugin to get pod ID by the application namespace

* Get rid of the uneccessary ContivRule ID.

* Order ContivRules by the size of the traffic matched.

* Renderer cache API - cache unified for both ACL and VPPTCP renderer

RendererCache combines capabilites of the VPPTCP and ACL caches
under a unified interface.

The rules are grouped into tables (ContivRuleTable type) and the
configuration is represented as a list of local tables, applied
on the ingress or the egress side of pods, and a single global table,
applied on the interfaces connecting the node with the rest
of the cluster.

The list of local tables is minimalistic in the sense that pods with
the same set of rules will share the same local table. Whether shared
tables are installed in one instance or as separate copies for each
associated pod is up to the renderer (usually determined by
the capabilities of the destination network stack).

All tables match only one side of the traffic - either ingress or
egress, depending on the cache orientation as selected in the Init method.
The cache combines the received ingress and egress Contiv rules
into the single chosen direction in a way that maintains the original
semantic (the global table is introduced to accomplish the task).

The rules are ordered in tables such that if rule *r1* matches subset
of the traffic matched by *r2*, then r1 precedes r2 in the list.
It is the order at which the rules should by applied by the rule
matching algorithm in the destination network stack (otherwise the
more specific rules could be overshadowed and never matched).

There are two types of tables distinguished:
  1. Local table: should be applied to match against traffic leaving
                  (IngressOrientation) or entering (EgressOrientation)
                  a selected subset of pods.
                  Every pod has at most one local table installed at
                  any given time. For a given local table, the set
                  of rules is immutable. Different content is treated
                  as a new local table (and the original table may
                  get unassigned from some or all originally
                  associated pods).
                  Local table has always at least one rule, otherwise
                  it is simply not tracked and returned by the cache.
  2. Global table: should be applied to match against traffic entering
                   (IngressOrientation) or leaving (EgressOrientation)
                   the node. There is always exactly one global table
                   installed (per node).
                   The global table may contain an empty set of rules
                   (meaning ALLOW-ALL).

* Refactor ACL Renderer to use the newly introduced RendererCache

* Refactor VPPTCP Renderer to use the newly introduced RendererCache

* RendererCache: implemented Resync, Commit, Changes and all getters

Update() method is still to-be-done.

* RendererCache: implemented Update() method.

* Policy Renderer: the IP of a removed pod may not be known.

In Resync we are not able to *easily* fully reconstruct the policy
configuration, most notably the IP addresses of pods.
For pods no longer existing after the resync it should not be
necessary to know the IP address anyway, therefore it can be nil.

* Commit forgotten file ports.go.

* RendererCache: fixed few bugs and golint issues.

* RendererCache: First UT.

* RendererCache: fixed few bugs + added helper methods for UTs.

* RendererCache: clean up of UTs.

* Integrate DHCP with SNAT.

* Update Contiv Mock with the newly added APIs.

* RendererCache: Finalize UTs.

* Fix VPPTCP renderer UTs.

* Fix golint issues.

* Mock for VPP/ACL engine.

* ACL renderer: UTs covering basic functionality

* ACL renderer: finalize UTs.

* containermap: Rename LookupPodNs to LookupPodAppNs and add UT.

* ACL renderer: add UT focused on sharing local tables.

* Remote CNI server: cover newly added methods with UTs.

* Fix policy cleanup after pod removal.

* Agent's fix temporarily added directly to vendor.

* Revert "Keep kube-system pods accessible regardless of the policies installed."

This reverts commit 20496443b9f2f8a5f9967205072b309c57cbca91.

Based on:
https://github.com/ahmetb/kubernetes-network-policy-recipes/blob/master/11-deny-egress-traffic-from-an-application.md
it is clear that policies should apply to the kube-system namespace
as normally.

* Policies should apply to kube-system namespace as well.

* Policy Processor: newly added pod may have IP address already assigned.

* Wait with policy resync until Node IP is set.

* Some extra logs for better debugging.

* Cleanup PolicyProcessor: Move repeated code to Process().

* Fix formatting.