docker/ubuntu-based: move vpp to separate image
This change makes it possible to build vpp and tag it with the commit.
This will enable us to reuse a built vpp Docker image. This image will
get rebuilt whenever the VPP git commit is changed.

This also drops some `--no-cache` arguments which don't make sense. We
should rely on Docker's invalidation and address caching issues another
way.

Signed-off-by: Cristian Staretu <unclejack@users.noreply.github.com>