DO NOT MERGE - ksr wiring POC
This commit is never intended to be merged.

Its a simple POC to try out wiring on ksr.

So far we have wiring for PluginName, Log, and PluginConfig
provided and testing for ksr.

There's more to be done.  This is a first step.

To do this, we introduce:

wiring/wiring_spi.go
wiring/wiring_impl.go

Neither of which in steady state belong here.

We also introduce:

plugin/ksr/wiring_impl.go
plugin/ksr/wiring_test.go

Signed-off-by: Ed Warnicke <hagbard@gmail.com>