Add arm64 support for Contiv/VPP docker images building
Add arm64 support for contiv/vpp docker images building
The naming rules of the built docker images is given below:
For a built image named as ${repo}:${tag}, or
${repo}/${component}:${tag},
1. For x86_64 arch, it would be named as:
${repo}-amd64:${tag}, or ${repo}/${component}-amd64:${tag}
2. For arm64 arch, it would be named as:
${repo}-arm64:${tag}, or ${repo}/${component}-arm64:${tag}
3. The docker images built for amd64 are tagged as the default:
${repo}-amd64:${tag}--->${repo}:${tag}
${repo}/${component}-amd64:${tag}--->${repo}/${component}:${tag}

The purpose of doing it in this way is to prepare the enablement
of the multi-arch support with fat manifest in a certain template
like:
 --template ${image}-ARCH:${tag}.
in the future.

In addition to build scripts and dockerfiles, the
corresponding utility scripts, such as push-all.sh,
save.sh are also updated to reflect the change.

This building process for this commit is tested through
both on arm64 and x86_64.

More detailed information about adding arm64 support for contiv/vpp
please refer:
https://github.com/contiv/vpp/issues/786

Signed-off-by: trevortao <trevor.tao@arm.com>