Fix the ordering of execution of vpp api trace commands
VPP API trace commands were a part of the VPP_COMMANDS associative
array, and the script loops through all elements of the array
when collecting VPP data. Unfortunately, the order in which
elements are accessed in the for loop is not guaranteed, so
the vpp dump trace command was executed before the vpp collect
trace command. This PR fixes the ordering by explicitly calling
the vpp api trace commands rather than executing them in a for
loop that iterates over the VPP_COMMANDS array.