Compress core dumps intermediately, use xz compression
Since the core dumps can be large, it may be possible to run out of
disk space on the machine running the bug report script.  Typically
it is not a good idea to compress compressed files, as it results
in a larger archive overall.  However, since this issue is being hit,
it's better to try and be more resilient.  The top level archive
compression is also changed to xz, for similar reasons.