Add yamllint to our travis-ci runs
This commit is inspired by a similar commit to Network Service Mesh [1], this
commit [2] from cn-infra, and this commit [3] to vpp-agent.

Since we are making heavy use of yaml files for Network Service Mesh, we should
ensure these are syntactically correct. We'll use yamllint [4] for this.

This commit adds running of yamllint into our travis-ci runs. It also adds a
custom configuration file which turns the yamllint line length check into a
warning. And finally, it fixes a large amount of yamllint errors in our
existing yaml files.

[1] https://github.com/ligato/networkservicemesh/pull/263
[2] https://github.com/ligato/cn-infra/pull/333
[3] https://github.com/ligato/vpp-agent/pull/820
[4] https://github.com/adrienverge/yamllint

Signed-off-by: Kyle Mestery <mestery@mestery.com>