Add multi-arch support for contivvpp images
Add multi-arch support for contivpp images by enabling
fat-manifest of images, now the supported platforms are
amd64 and arm64.

With the support, a simple image name as contivvpp/${image}:${tag}
can be used in pulling the right images for the matched x86_64 or
aarch64 platform.
It would also simply the deployment of Contiv/VPP in k8s environment.

Signed-off-by: Trevor Tao <trevor.tao@arm.com>